Implement Dictionary Batch
