+++ transformer/data/create_character_map.py
@@ -1,8 +1,6 @@
-from typing import List, Set, Dict, Tuple, Optional
 from collections import Counter
 import os
 
-import torch
 from datasets import load_dataset
