+++ transformer/modules/transformer_modules.py
@@ -4,7 +4,8 @@ import math
 
 import torch
 
-from HobBeans.transformer.config_base import UnionLikeConfig, BaseConfig
+from transformer.configs.config_base import UnionLikeConfig, BaseConfig
+from transformer.data.batch_base import DictionaryBatch
 from matplotlib import pyplot as plt
 
 
@@ -275,7 +276,7 @@ class EncoderBlock(torch.nn.Module):
     return x
 
 
-class AutoregressiveTransformerEncoderConfig(BaseConfig):
+class AutoregressiveTransformerConfig(BaseConfig):
   """
   Config for the autoregressive transformer encoder.
 
@@ -292,13 +293,13 @@ class AutoregressiveTransformerEncoderConfig(BaseConfig):
   num_layers: int
 
 
-class AutoregressiveTransformerEncoder(torch.nn.Module):
+class AutoregressiveTransformer(torch.nn.Module):
   """
-  Autoregressive transformer encoder.
+  Autoregressive transformer decoder.
   """
 
   def __init__(
-    self, config: AutoregressiveTransformerEncoderConfig, pad_token: Optional[int], vocab_size: int
+    self, config: AutoregressiveTransformerConfig, pad_token: Optional[int], vocab_size: int
   ):
     super().__init__()
     self.pad_token = pad_token
@@ -330,6 +331,21 @@ class AutoregressiveTransformerEncoder(torch.nn.Module):
       x = encoder_block(x, key_padding_mask=key_padding_mask, is_causal=False, attn_mask=attn_mask)
     x = self.output_projection(x)
     return x
+
+
+class TransformerWrapper(torch.nn.Module):
+  def __init__(self, *, config: AutoregressiveTransformerConfig, pad_token: int, vocab_size: int):
+    super().__init__()
+    self.config = config
+    self.pad_token = pad_token
+    self.vocab_size = vocab_size
+    self.model = AutoregressiveTransformer(config, pad_token, vocab_size)
+
+  def forward(self, batch: DictionaryBatch) -> DictionaryBatch:
+    result = DictionaryBatch()
+    x = batch["inputs"]
+    result["logits"] = self.model(x)
+    return result
 
 
 if __name__ == "__main__":
